initiating interpreter class implementation

+++ ToyLanguage/ToyLanguage/Interpreter/Exceptions/InvalidNumberOfArgumentsException.cs
@@ -4,7 +4,7 @@ namespace ToyLanguage.Interpreter.Exceptions
 {
     internal class InvalidNumberOfArgumentsException : Exception
     {
-        public InvalidNumberOfArgumentsException(string functionName, int expected, int got) : base($"<{functionName}> expected {expected} arguments , got {got}`")
+        public InvalidNumberOfArgumentsException(string functionName, double expected, int got) : base($"<{functionName}> expected {expected} arguments , got {got}`")
         {
         }
     }
